+++ Monty/src/main/java/sml/casts/FloatToInt.java
@@ -16,13 +16,13 @@ limitations under the License.
 
 package sml.casts;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class FloatToInt {
 
-	public static BigInteger floatToInt(Float floating) {
+	public static BigInteger floatToInt(BigDecimal floating) {
 		return new BigInteger(floating.toString());
-
 	}
 
 }
